Rename payment.py to invoice.py.
