+++ hypha/apply/dashboard/tests/test_views.py
@@ -6,7 +6,7 @@ from hypha.apply.funds.tests.factories import (
     ApplicationSubmissionFactory,
     InvitedToProposalFactory,
 )
-from hypha.apply.projects.models.payment import (
+from hypha.apply.projects.models.invoice import (
     APPROVED_BY_STAFF,
     CHANGES_REQUESTED_BY_STAFF,
     DECLINED,
