+++ hypha/apply/projects/tests/test_more_templatetags.py
@@ -11,7 +11,7 @@ from hypha.apply.users.tests.factories import (
     StaffFactory,
 )
 
-from ..models.payment import (
+from ..models.invoice import (
     DECLINED,
     PAID,
     SUBMITTED,
