+++ hypha/apply/projects/utils.py
@@ -16,7 +16,7 @@ from .constants import (
     INT_VENDOR_PENDING,
 )
 from .models import Project
-from .models.payment import (
+from .models.invoice import (
     APPROVED_BY_FINANCE,
     APPROVED_BY_STAFF,
     CHANGES_REQUESTED_BY_FINANCE,
