+++ hypha/apply/projects/views/invoice.py
@@ -68,7 +68,7 @@ from ..forms import (
     EditInvoiceForm,
     InvoiceTagsForm,
 )
-from ..models.payment import (
+from ..models.invoice import (
     APPROVED_BY_FINANCE,
     APPROVED_BY_STAFF,
     CHANGES_REQUESTED_BY_FINANCE,
